stash

+++ onnxruntime/contrib_ops/cpu/bert/bias_gelu_fusion.cc
@@ -52,9 +52,8 @@ Status BiasGelu<T>::Compute(OpKernelContext* ctx) const {
   T* Y_data = Y->template MutableData<T>();
   int64_t task_count = X->Shape().Size() / bias_len;
 
-  concurrency::ThreadPool::TryBatchParallelFor(ctx->GetOperatorThreadPool(),
-                                               static_cast<int32_t>(task_count),
-                                               [&](int32_t task_idx) {
+  concurrency::ThreadPool::TryBatchParallelFor(ctx->GetOperatorThreadPool(), static_cast<int32_t>(task_count),
+                                               [&](ptrdiff_t task_idx) {
                                                  const T* p_input = X_data + task_idx * bias_len;
                                                  T* p_output = Y_data + task_idx * bias_len;
                                                  T* p_output_tmp = tmp_data + task_idx * bias_len;
